requirements.txt
git+https://github.com/nledez/python-certifi-debian
python-consul
pyyaml
hvac
